add data-api support for log statistics
