+++ components/_util/getDataAttr.tsx
@@ -1,0 +1,9 @@
+export default function getDataAttr(props) {
+  const dataAttrs = {};
+  Object.keys(props).forEach(i => {
+    if (i.indexOf('data-') === 0) {
+      dataAttrs[i] = props[i];
+    }
+  });
+  return dataAttrs;
+}
