+++ components/input-item/demo/1.md
@@ -17,6 +17,7 @@ let BasicInputExample = React.createClass({
       <InputItem
         defaultValue=""
         placeholder="设置defaultValue,不设置value"
+        data-seed="logId"
       >非受控</InputItem>
       <InputItem
         {...getFieldProps('control', {
