+++ components/popover/demo/basic.md
@@ -34,7 +34,7 @@ const App = React.createClass({
         <Popover
           visible={this.state.visible}
           overlay={[
-            (<Item key="4" value="scan" iconName="scan">扫一扫</Item>),
+            (<Item key="4" value="scan" iconName="scan" data-seed="logId">扫一扫</Item>),
             (<Item key="5" value="special" iconName="qrcode" style={{ whiteSpace: 'nowrap' }}>我的二维码</Item>),
             (<Item key="6" value="button ct" iconName="question-circle-o">
               <span style={{ marginRight: 5 }}>帮助</span>
