+++ components/textarea-item/demo/1.md
@@ -17,6 +17,7 @@ let TextareaItemExample = React.createClass({
         title="非受控"
         defaultValue=""
         placeholder="设置defaultValue,不设置value"
+        data-seed="logId"
       />
       <TextareaItem
         title="受控组件"
